requirements.txt
configparser
psycopg2-binary
click
python-dotenv
nltk
pycountry